Cleanly handle case where we exit before target boots

+++ scripts/runbenchmark.sh
@@ -79,24 +79,28 @@ clean_benchmark()
   error=$?
   local clean=0
 
-  if test x"${target_dir}" = x; then
-    echo "No directory to remove from ${ip}" 1>&2
-  elif test x"${keep}" = '-k'; then
-    echo "Not removing ${target_dir} from ${ip} as -k was given. You might want to go in and clean up."
-    clean=1
-  elif ! expr "${target_dir}" : '\(/tmp\)' > /dev/null; then
-    echo "Cowardly refusing to delete ${target_dir} from ${ip}. Not rooted at /tmp. You might want to go in and clean up." 1>&2
-    error=1
-    clean=1
-  else
-    (. "${topdir}"/lib/common.sh; remote_exec "${ip}" "rm -rf ${target_dir}")
-    if test $? -eq 0; then
-      echo "Removed ${target_dir} from ${ip}"
-    else
-      echo "Failed to remove ${target_dir} from ${ip}. You might want to go in and clean up." 1>&2
+  if test x"${ip}" != x; then
+    if test x"${target_dir}" = x; then
+      echo "No directory to remove from ${ip}" 1>&2
+    elif test x"${keep}" = '-k'; then
+      echo "Not removing ${target_dir} from ${ip} as -k was given. You might want to go in and clean up."
+      clean=1
+    elif ! expr "${target_dir}" : '\(/tmp\)' > /dev/null; then
+      echo "Cowardly refusing to delete ${target_dir} from ${ip}. Not rooted at /tmp. You might want to go in and clean up." 1>&2
       error=1
       clean=1
+    else
+      (. "${topdir}"/lib/common.sh; remote_exec "${ip}" "rm -rf ${target_dir}")
+      if test $? -eq 0; then
+        echo "Removed ${target_dir} from ${ip}"
+      else
+        echo "Failed to remove ${target_dir} from ${ip}. You might want to go in and clean up." 1>&2
+        error=1
+        clean=1
+      fi
     fi
+  else
+    echo "Target post-boot initialisation did not happen, thus nothing to clean up."
   fi
 
   if test x"${listener_pid}" != x; then
